Version 1.0
This is the first major update to the site. It includes a move away from the GitHub Pages hosted Jekyll CV and into a Hugo Extended based page, built on the [Wowchemy](https://wowchemy.com/) Academic theme. The page is built with GitLab Continuous Integrations (GitLab CI) pipelines and deployed to Google Cloud Platform Firebase.

This version include a basic introduction, self-professed skills, and work history.